Development version 1.1-dev including the following changes:
* Added the ability to create and manage sub-groups.
* In the "Display my favorite contacts" topic, the contact's informations are displayed in a text control, to facilitate copying.
* Under "My favorite contacts", the values entered for each contact are multiline.

+++ addon/installTasks.py
@@ -7,4 +7,5 @@ import globalVars
 CONFIG_FILE_NAME = "myFavoriteTopics.ini"
 
 def onInstall():
-	shutil.move(os.path.join(os.path.dirname(__file__), CONFIG_FILE_NAME), os.path.join(globalVars.appArgs.configPath, CONFIG_FILE_NAME))
+	if not os.path.exists(os.path.join(globalVars.appArgs.configPath, CONFIG_FILE_NAME)):
+		shutil.move(os.path.join(os.path.dirname(__file__), CONFIG_FILE_NAME), os.path.join(globalVars.appArgs.configPath, CONFIG_FILE_NAME))
